rename toaster, refactor the switch
