+++ examples/react-client/fishjam-chat/src/main.tsx
@@ -5,8 +5,8 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 
 import App from "./App.tsx";
-import { BlurProvider } from "./components/BlurToggle.tsx";
-import { Toaster } from "./components/ui/sonner.tsx";
+import { BlurProvider } from "./components/BlurToggle";
+import { Toaster } from "./components/ui/toaster";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
